Перенос моделей и моделей-представленя в проект Domain
