+++ Common/WebStore.Domain/ViewModels/CartOrderViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace WebStore.ViewModels
+﻿namespace WebStore.Domain.ViewModels
 {
     public class CartOrderViewModel
     {
