+++ Common/WebStore.Domain/ViewModels/CatalogViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
     public class CatalogViewModel
     {
