+++ Common/WebStore.Domain/ViewModels/EmployeeViewModel.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNetCore.Mvc;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
     public class EmployeeViewModel : IValidatableObject
     {
