+++ Common/WebStore.Domain/ViewModels/ProductViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace WebStore.ViewModels
+﻿namespace WebStore.Domain.ViewModels
 {
     public class ProductViewModel
     {
