+++ Common/WebStore.Domain/ViewModels/SectionViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace WebStore.ViewModels
+namespace WebStore.Domain.ViewModels
 {
     public class SectionViewModel
     {
