+++ UI/WebStore/Data/TestData.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using WebStore.Domain.Entities;
-using WebStore.Models;
+using WebStore.Domain.Models;
 
 namespace WebStore.Data
 {
