+++ UI/WebStore/Services/InCookies/InCookiesCartService.cs
@@ -7,9 +7,9 @@ using Microsoft.AspNetCore.Http;
 using Newtonsoft.Json;
 
 using WebStore.Domain.Entities;
+using WebStore.Domain.ViewModels;
 using WebStore.Infrastructure.Mapping;
 using WebStore.Services.Interfaces;
-using WebStore.ViewModels;
 
 namespace WebStore.Services.InCookies
 {
