+++ UI/WebStore/Services/InMemory/InMemoryEmployeesData.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Extensions.Logging;
 using WebStore.Data;
-using WebStore.Models;
+using WebStore.Domain.Models;
 using WebStore.Services.Interfaces;
 
 namespace WebStore.Services.InMemory
