+++ UI/WebStore/Services/Interfaces/ICartService.cs
@@ -1,4 +1,4 @@
-﻿using WebStore.ViewModels;
+﻿using WebStore.Domain.ViewModels;
 
 namespace WebStore.Services.Interfaces
 {
