+++ UI/WebStore/Services/Interfaces/IEmployeesData.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using WebStore.Models;
+using WebStore.Domain.Models;
 
 namespace WebStore.Services.Interfaces
 {
