+++ UI/WebStore/Views/Account/Login.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebStore.ViewModels.Identity.LoginViewModel
+﻿@model WebStore.Domain.ViewModels.Identity.LoginViewModel
 @{
     ViewData["Title"] = "Вход в систему";
 }
