+++ UI/WebStore/Views/Account/Register.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebStore.ViewModels.Identity.RegisterUserViewModel
+﻿@model WebStore.Domain.ViewModels.Identity.RegisterUserViewModel
 @{
     ViewData["Title"] = "Register";
 }
