+++ UI/WebStore/Views/Cart/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model CartOrderViewModel
+﻿@model WebStore.Domain.ViewModels.CartOrderViewModel
 @{
     ViewBag.Title = "Корзина";
 }
