+++ UI/WebStore/Views/Cart/Partial/_CheckOut.cshtml
@@ -1,4 +1,4 @@
-﻿@model OrderViewModel
+﻿@model WebStore.Domain.ViewModels.OrderViewModel
 
 <form asp-action="CheckOut" asp-antiforgery="true">
     <div class="row">
