+++ UI/WebStore/Views/Catalog/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model CatalogViewModel
+﻿@model WebStore.Domain.ViewModels.CatalogViewModel
 @{
     ViewBag.Title = "Каталог товаров";
 }
