+++ UI/WebStore/Views/Employees/Delete.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebStore.ViewModels.EmployeeViewModel
+﻿@model WebStore.Domain.ViewModels.EmployeeViewModel
 
 @{
     ViewData["Title"] = "Delete";
