+++ UI/WebStore/Views/Employees/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebStore.Models.Employee
+﻿@model WebStore.Domain.Models.Employee
 
 @{
     ViewData["Title"] = "Details";
