+++ UI/WebStore/Views/Employees/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebStore.ViewModels.EmployeeViewModel
+﻿@model WebStore.Domain.ViewModels.EmployeeViewModel
 
 @{
     ViewData["Title"] = "Edit";
