+++ UI/WebStore/Views/Employees/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@using WebStore.Domain.Entities.Identity
-@model IEnumerable<WebStore.Models.Employee>
+@model IEnumerable<WebStore.Domain.Models.Employee>
 
 <div class="container">
     <h3>Сотрудники</h3>
