+++ UI/WebStore/Views/UserProfile/Orders.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<UserOrderViewModel>
+﻿@model IEnumerable<WebStore.Domain.ViewModels.UserOrderViewModel>
 @{
     ViewBag.Title = "Профиль пользователя";
     Layout = "_UserProfileLayout";
