+++ UI/WebStore/Views/_ViewImports.cshtml
@@ -1,6 +1,4 @@
 ﻿@using WebStore
-@using WebStore.Models
-@using WebStore.ViewModels
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, WebStore
